added 'skippy' video mode which draws every other video line, and gives about 50% of the cpu back to the main program
TODO: still need to add in ps2keyboard polling.

idea for polling:  the original cat1 demo driver sampled at around hsync (the same time each line), and then if the sample showed a change, at the END of the scanline, it handled the key.  I did this in the blanking interval instead of returning to the main application.  I may have to do that here.  If the change happened on a blank or skippy line, there's no problem.

For now, if it shows a change, lets just aboirt the current line.  it'll disrupt the display, but whatever.

Also might be able to just 'notch' the display off the side.  The atari 2600 does this all the time.